Moved dendrogram tests to optional
